+++ app/(pages)/_data/navLinks.json
@@ -17,6 +17,6 @@
     },
     {
       "name": "Content Collection",
-      "slug": "/content/content"
+      "slug": "/content-collection"
     }
   ]
