+++ app/(pages)/content-collection/_component/CollectionCard/CollectionCard.tsx
