+++ app/(pages)/content-collection/_component/CollectionCard/MediaCard.tsx
