+++ app/(pages)/content-collection/_component/CollectionCard/MediaPreview.tsx
