+++ app/(pages)/content-collection/_component/CollectionCard/PreviewCard.tsx
